chore: tidying logic
